+++ designsystem/src/main/res/drawable/ico_eco.xml
