+++ playground/src/main/java/com/everli/designsystem/playground/theme/TokensPlayground.kt
@@ -19,7 +19,7 @@ fun TokensPlayground() {
     horizontalAlignment = Alignment.CenterHorizontally,
     modifier = Modifier
       .fillMaxHeight()
-      .fillMaxWidth()
+      .fillMaxWidth(),
   ) {
     Text("\uD83D\uDEA7 Work in progress \uD83D\uDEA7", modifier = Modifier)
   }
