--- playground/src/main/res/layout/layout_button_sample.xml
@@ -1,94 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<ScrollView
-  xmlns:android="http://schemas.android.com/apk/res/android"
-  xmlns:app="http://schemas.android.com/apk/res-auto"
-  xmlns:tools="http://schemas.android.com/tools"
-  android:layout_width="match_parent"
-  android:layout_height="match_parent">
-
-  <LinearLayout
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:orientation="vertical"
-    tools:ignore="HardcodedText">
-
-    <com.everli.designsystem.components.button.EverliButtonView
-      android:layout_width="match_parent"
-      android:layout_height="wrap_content"
-      android:layout_marginHorizontal="12dp"
-      app:btnIcon="@drawable/ico_add_circle"
-      app:btnSize="large"
-      app:btnStyle="fill"
-      app:btnVariant="primary"
-      app:text="Primary Fill with Icon"
-      app:useContextTheme="true" />
-
-    <com.everli.designsystem.components.button.EverliButtonView
-      android:layout_width="match_parent"
-      android:layout_height="wrap_content"
-      android:layout_marginHorizontal="12dp"
-      app:btnSize="medium"
-      app:btnStyle="outline"
-      app:btnVariant="primary"
-      app:text="Primary Outline"
-      app:useContextTheme="true" />
-
-    <com.everli.designsystem.components.button.EverliButtonView
-      android:layout_width="match_parent"
-      android:layout_height="wrap_content"
-      android:layout_marginHorizontal="12dp"
-      app:btnIcon="@drawable/ico_add"
-      app:btnSize="medium"
-      app:btnStyle="flat"
-      app:btnVariant="primary"
-      app:iconPosition="right"
-      app:text="Flat Icon Right"
-      app:useContextTheme="true" />
-
-    <com.everli.designsystem.components.button.EverliButtonView
-      android:layout_width="wrap_content"
-      android:layout_height="wrap_content"
-      android:layout_marginHorizontal="12dp"
-      app:btnSize="large"
-      app:btnVariant="primary"
-      app:text="Default Theme" />
-
-    <com.everli.designsystem.components.button.EverliButtonView
-      android:layout_width="wrap_content"
-      android:layout_height="wrap_content"
-      android:layout_marginHorizontal="12dp"
-      app:btnIcon="@drawable/logo_facebook"
-      app:btnSize="large"
-      app:btnVariant="facebook"
-      app:text="Default Theme" />
-
-    <com.everli.designsystem.components.button.EverliButtonView
-      android:layout_width="wrap_content"
-      android:layout_height="wrap_content"
-      android:layout_marginHorizontal="12dp"
-      app:btnIcon="@drawable/logo_google"
-      app:btnSize="medium"
-      app:btnVariant="google"
-      app:text="Default Theme" />
-
-    <com.everli.designsystem.components.button.EverliButtonView
-      android:layout_width="wrap_content"
-      android:layout_height="wrap_content"
-      android:layout_marginHorizontal="12dp"
-      app:btnIcon="@drawable/logo_payment_blik"
-      app:btnSize="medium"
-      app:btnVariant="blik" />
-
-    <com.everli.designsystem.components.button.EverliButtonView
-      android:layout_width="wrap_content"
-      android:layout_height="wrap_content"
-      android:layout_marginHorizontal="12dp"
-      app:btnIcon="@drawable/logo_apple"
-      app:btnSize="medium"
-      app:btnVariant="apple"
-      app:text="Disabled" />
-
-  </LinearLayout>
-
-</ScrollView>
-
