+++ playground/src/main/res/values/themes.xml
@@ -12,13 +12,6 @@
     <!-- Status bar color. -->
     <item name="android:statusBarColor" tools:targetApi="l">?attr/colorPrimaryVariant</item>
     <!-- Customize your material theme here. -->
-
-    <!-- Design system specific -->
-    <item name="buttonColorPrimaryFillBackgroundEnabled">@color/purple_100</item>
-    <item name="buttonColorPrimaryFillBackgroundPressed">@color/purple_80</item>
-    <item name="buttonColorPrimaryOutlineBackgroundPressed">@color/purple_20</item>
-    <item name="buttonColorPrimaryOutlineBorderEnabled">@color/purple_80</item>
-    <item name="buttonColorPrimaryFlatBackgroundPressed">@color/purple_20</item>
   </style>
 
   <style name="Theme.DesignSystemAndroid.NoActionBar">
